Enhance cadvisor and kube-kubelet scrape job configs
Restricting the service discovery for cadvisor and kube-kubelet to the
`kube-system` namespace helps reduce the number of potential targets
Prometheus has to discard during relabeling.

Setting `follow_redirects` to false prevents Prometheus from being
redirected to scrape a different target in the seed if an attacker were to
tweak a scraped target. This could otherwise lead to information leakage.