Automatically enable admission controller for provider extensions in local extension setup. (#8311)
* Automatically enable admission controller for provider extensions in local extension setup.

Some of the validation logic is only available in the admission controllers of provider
extensions. Therefore, it is helpful if those admission controllers are installed in the
local extension setup as well.
This change attempts to identify the latest admission controller matching to the provider
extension in use. The version might not match, but admission controllers are fairly stable.
In case a provider extension from a non-gardener repository is used no attempt at
discovering its repository is tried.

* Address review feedback