Fix the average over time calculation (!)
Initially, the fallback case

    or
      ...:sum_over_time

was added for the first iteration, for the initialization phase when the time
was still zero and hence the ratio could not be calculated because it would lead
to a division by zero.

This turned out to be a major _bug_: first of all, it is not needed in the first
iteration and in the finalization phase, it could lead to completely bogus
average values when the sum was used as average, being orders of magnitude more
than the actual average value.

When both series exist but the time is zero, we could set the average
value to zero like this:

        metering:persistent_volume_claims:sum_by_namespace:sum_over_time * 60
      /
        (metering:disk_usage_seconds != 0)
    or
        (metering:disk_usage_seconds == 0)
      +
        metering:persistent_volume_claims:sum_by_namespace:sum_over_time * 0

However, having only the time series as 0 and no avg values is just as good in
the end and less confusing to the reader.