Update the copy of the stateful metering rules of the aggregate Prometheus
This file used to be a symbolic link. "//go:embed" does not support symbolic
links hence this file is now a real copy of the stateful metering recording
rules of the "cache" Prometheus. The changes of the previous commits are now
reflected in this copy as well.

See PR 8284.

Background on the rational of this duplicated file:

The aggregate Prometheus instances independently execute the same stateful
recording rules as the cache Prometheus instances, rather than federating them
from the cache Prometheus. This approach minimizes network traffic, albeit at
the expense of increased computational overhead.

The rationale behind this trade-off is that by federating only "stateless"
metrics from the cache Prometheus, we can operate multiple cache Prometheus
instances and federate from any one of them. In a similar vein, we can run
multiple aggregate Prometheus instances and federate the stateless metrics from
any one of them, even though this means calculating the stateful metrics across
all aggregation levels.

It's important to note that the stateful metrics include the "carry-on" time
series :this_month, which is utilized to circumvent the need to comb through a
month's worth of data points to obtain the resource usage metrics for shoots
that have already been deleted. If there's a high turnover rate of shoots,
federating these "carry-on" metrics could result in substantial network usage.
However, maintaining them is computationally less costly as they hold a constant
value.