`care` operations package refactoring (#8387)
* Extract `garden` health

Move all `garden` specific health checks to the respective `care` controller.

* Extract `seed` health

Move all `seed` related checks to the respective `care` controller.

* Move shoot specific checks to `health`

* Move `HealthChecker` to separate package

* Move `CheckClusterNodes` to shoot care

This functionality is very specific to shoot clusters and should therefore reside in the shoot care related package.

* Move `ComputeTechnicalID` to `gardenerutils`

This eliminates further "odd" dependencies for `gardener-operator`.

* Move `NewConditionOrError` to `v1beta1helper`

* Fix nil pointer deref

* Move `FailedCondition` to `v1beta1helper`

* move condition related helpers to one file

* Reduce fields in `HealthChecker`

* Drop `checkManagedResourceFor{Seed,Garden}`

* Move `TestingShoot` case to controller logic

* Remove obsolete tests

- `#PardonCondition` test is part of `pkg/operation/care/health_test.go`
- `#CheckLoggingControlPlane` test is part of `pkg/utils/kubernetes/health/checker/checker_test.go`
- The Testing shoot use case is removed since it became part of `checkObservabilityComponents` in `pkg/operation/care/health.go`

* Add `healthChecker` field to `Health` struct

This helps to reduce function arguments

* Add tests for `ComputeRequired*` functions

* Remove `shootClient` from struct

The `shootClient` is not always available/initialized and thus it's critical for functions to rely on this struct field. Hence, it's better to pass it via arguments to the functions which need the client.

* Remove code duplication in `operation/care/health`

* Address review feedback

* Use condition/constraint structs

Earlier, condition and constraint variable handling was scattered. Using dedicated structs help to reduce the involved complexity and code duplication.

* Address review comments

* Drop `pkg/operation/care` package