Refactor plutono dashboard configmap unit test
Previously, the hard-coded hashes in this test required the developer to
change the hash every time they made a change. The only way to figure out
what the new hash would be was to start the debugger, find the name of the
configmap, and to copy and paste that into the code. That is cumbersome to
work around.
Now, the ConfigMaps are retrieved and matched against a substring of their
name. This retains most of the information and is reasonably robust for a
test.