Add failing tests
They fail because the `ControllerRegistration` still exists after the `Eventually` timeout in L469 because of

```
"error":"cannot remove finalizer of ControllerRegistration \"ctrlreg-controllerregistration-controller-test-wd67p\" because still found ControllerInstallations: [ctrlreg-controllerregistration-controller-test-wd67p-gwrsf]"
```

This is exactly the bug. The `ControllerRegistration` controller does not delete the `ControllerInstallation`s when the deployment policy is `Always` or `AlwaysExceptNoShoots`.