`Garden` controller waits for `etcd-druid` deployment
Otherwise, it might proceed with reconciliation of the `Etcd` resources even if `etcd-druid` is still in the phase of getting rolled out/updated. This can be especially dangerous/problematic when the images in the `etcd-druid`s image vector change. Without this, the configuration in the `Etcd` resource might not match the images or the `etcd-druid` version, leading to issues with the resulting ETCD `StatefulSet`s.