`Garden` controller waits for `etcd-druid` deployment (#8456)
* `Garden` controller waits for `etcd-druid` deployment

Otherwise, it might proceed with reconciliation of the `Etcd` resources even if `etcd-druid` is still in the phase of getting rolled out/updated. This can be especially dangerous/problematic when the images in the `etcd-druid`s image vector change. Without this, the configuration in the `Etcd` resource might not match the images or the `etcd-druid` version, leading to issues with the resulting ETCD `StatefulSet`s.

* Make "wait until healthy and not progressing" function reusable

Note: Unit tests are copied from the `WaitUntilHealthy` function. It is not a goal of this PR to improve these unit tests.

* Adapt `Wait` of `etcd-druid` to wait until `ManagedResource` is no longer progressing