Increase the look back time from 10 minutes to 30 minutes
The last_over_time(...[10m]) pattern was used in the "continuation" series to
mitigate the case when a Prometheus restart takes more than 5 minutes: time
series get stale after 5 minutes by default. The default staleness behavior of
an instant vector was replaced with checking the data points in the last 10
minutes using a range vector, thereby increasing the "staleness" period from 5
minutes to 10 minutes.

It turns out that due to slow disk attachment and WAL replays, a Prometheus
restart can take about 10 minutes. This commit hence further increases the
staleness period to 30 minutes. Note that Prometheus restarts are expected to be
infrequent and fast in general.