[operator] Introduce `garden-reference` controller (#8439)
* Prefactor: Remove code duplication (1)

Merge `getUnreferencedSecrets` and `getUnreferencedConfigMaps` into `getUnreferencedObjects`

* Prefactor: Remove code duplication (2)

Introduce `releaseUnreferencedResources`

* Prefactor: Remove code duplication (3)

Introduce `getReferencedSecretNames` and `getReferencedConfigMapNames`

* Prefactor: Remove code duplication (4)

Merge `handleReferencedSecrets` and `handleReferencedConfigMaps` into `handleReferencedResources`

* Prefactor: Inline functions that are only called once

* Prefactor: Don't call `getReferenced*Names` twice

* Prefactor: Group code together for better readability

Just reordering, no functional changes whatsoever

* Drop broken gomock-based reconciler unit tests

Meanwhile, we have an envtest-based integration test for the controller, so the unit tests could be considered duplication, and adapting these gomock-based tests now is pointless (we actually don't want to use gomock anymore, so either we have to drop the tests or rewrite with a fake client, though the latter is again useless due to the already existing envtest-based integration test).

* ------------ empty separator commit -------------

* Make `shoot-reference` controller reusable

* Make `namespace` optional in `reference` controller

`Shoot`s are namespaced, but `Garden`s are not, so we have to provide a way to tell in which namespace the referenced objects are expected.

* ------------ empty separator commit -------------

* Drop useless `namespace` from ETCD backup `secretRef`

This was never working/used anyways, so let's drop it from the API

* Add documentation

* Add `garden-reference` controller

* Add integration test

* Address PR review feedback (nits)

* Forbid adding well-known finalizers

* Enhance integration test with multiple objects referencing the same secrets/configmaps

* Fix bug in logic

* Address PR review feedback