Don't use the red color if there are more than 80 shoots in a seed
Previously, a red color was used for seeds with more than 80 shoots.

Instead of this staticly configured limit, the seed's capacity should be
used. However, in Grafana / Plutono, the threshold feature only supports
static values, so we can not use that directly instead of the static value.

In a later PR we shall explore how could we show the seed's actual
capacity maybe with a different visualisation. The table visualisation
could be used to show both the actualy usage and the seed's capacity.