Sort the entries in the bar gauge panel
Note that it does not seem to be possible to sort the entries in Plutono
by the value of the "seed" label, and they are not returned in a sorted
order by Prometheus by default.

Hence the entries are kind of randomly sorted.

With this commit, we can sort them in descending order (showing the seeds
with most of the shoots on top).