Introduce `main()` function for `gardener-node-agent`
If it will be started with a kubeconfig pointing to the well-known location of the bootstrap token file, it will use it to fetch the access token and write it to the disk. Afterwards, it will overwrite the kubeconfig to point to the access token file instead of the bootstrap token file. The rest config will be modified ad-hoc, hence, the bootstrap token will only be used for fetching the "real" access token.