[node-agent] Introduce `cmd/gardener-node-agent/app.go` with `manager.Manager` initialization (#8627)
* Add `Server` configuration to `node-agent` config API

* Introduce `main()` function for `gardener-node-agent`

If it will be started with a kubeconfig pointing to the well-known location of the bootstrap token file, it will use it to fetch the access token and write it to the disk. Afterwards, it will overwrite the kubeconfig to point to the access token file instead of the bootstrap token file. The rest config will be modified ad-hoc, hence, the bootstrap token will only be used for fetching the "real" access token.

* Drop validation regarding kubeconfig

Otherwise, it's not possible to specify the kubeconfig via the `KUBECONFIG` environment variable

* Add GNA to `Dockerfile` and `Makefile`

* Address PR review feedback