Slightly improve the "Shoots per Seed" panel of the "Seed Resource Usage" dashboard (#8626)
* Update the dashboard's serialisation by exporting it unchanged

This commit contains only changes that are related to the differences
in how older or slightly newer versions of Grafana or Plutono serialize
dashboards.

  -  "schemaVersion": 25,
  +  "schemaVersion": 27,

  - 7.0.3
  + 7.5.24

* Remove the max 100 limit on the Shoots per Seed panel

If a seed had more than 100 shoots, the visualisation was chopped off
at 100, which was confusing.

* Don't use the red color if there are more than 80 shoots in a seed

Previously, a red color was used for seeds with more than 80 shoots.

Instead of this staticly configured limit, the seed's capacity should be
used. However, in Grafana / Plutono, the threshold feature only supports
static values, so we can not use that directly instead of the static value.

In a later PR we shall explore how could we show the seed's actual
capacity maybe with a different visualisation. The table visualisation
could be used to show both the actualy usage and the seed's capacity.

* Sort the entries in the bar gauge panel

Note that it does not seem to be possible to sort the entries in Plutono
by the value of the "seed" label, and they are not returned in a sorted
order by Prometheus by default.

Hence the entries are kind of randomly sorted.

With this commit, we can sort them in descending order (showing the seeds
with most of the shoots on top).