Improved alerts for ETCD backup failures. (#8355)
* Formats the snapshots alerts for better readablility.

* Improved alerts for ETCD backup failures.

Previously, the expression checked that the last snapshot is older than
a specified time, and a full or incremental backup is required, and
the etcd instance is a leader. It checked this by using arithmetics and
bool comparison expressions. In this commit,  it is performing an equivalent check
by cartessian product of  all the values and checking if the result is
more than 0.

Example: The first part `((time() -
etcdbr_snapshot_latest_timestamp{job="kube-etcd3-backup-restore-main",kind="Incr"}
> bool 900)` returns a set of ETCD instances, if metrics `etcdbr_snapshot_latest_timestamp`
with `kind="Incr"` is older than 900 seconds.
The second part
`etcdbr_snapshot_required{job="kube-etcd3-backup-restore-main",kind="Incr"})`
returns a set of ETCD instances if metrics `etcdbr_snapshot_required`
with `kind="Incr"` is set.
The third part `on (pod, role)
etcd_server_is_leader{job="kube-etcd3-main"}` returns the leader of ETCD instances on pod and role.
If the cartessian product of the three parts is greater than 0 then an
alert is raised. In plain English word, if the leader of ETCD instance
requires an incremental snapshot to be taken  and has
last incremental snapshot that is more than 900 seconds old, then an
alert is raised.