Add dns records for alertmanager to perform migration from ingress to istio without major disruption.
In the past, alertmanager was exposed via a wildcard dns entry pointing to the ingress, i.e. nginx.
Now, for the migration the first step is to create a specific dns record for each alertmanager instance
upon reconciliation and expose it via istio. Once all alertmanager instances have migrated to istio,
the specific dns records can be removed again and the wildcard dns entry can be created pointing
to istio.
As this is only a temporary migration process, the dns records were already annotated so that they
can be simply removed from the managed resource once the migration is finished.