Add dns records for vali to perform migration from ingress to istio without major disruption.
In the past, vali was exposed via a wildcard dns entry pointing to the ingress, i.e. nginx.
Now, for the migration the first step is to create a specific dns record for each vali instance
upon reconciliation and expose it via istio. Once all vali instances have migrated to istio,
the specific dns records can be removed again and the wildcard dns entry can be created pointing
to istio.
As this is only a temporary migration process, the dns records were already annotated so that they
can be simply removed from the managed resource once the migration is finished.