[node-agent] Introduce `gardener-node-init` script and unit (#8726)
* Add missing defaulting to GNA config

- client connection
- servers
- logging settings

* Adapt `UnitsToDiskScript` to consider new `.spec.units.files[]` field

* Restart GNA unit only after last-applied OSC has been persisted

Otherwise, GNA would restart itself too early, effectively triggering an infinite loop because it would always consider itself as "changed unit".
We already followed the same approach in GNA's `node` controller.

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* [prefactor] Extract common run command initialisation

This code is the very same for almost each binary, so let's extract it into a function to not duplicate it again and again.

* Introduce bootstrap command

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Introduce `gardener-node-agent` component in original OSC package

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Introduce `gardener-node-init` script and unit

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Address PR review feedback (graceful shutdown)

- OSC controller calls the root `context.CancelFunc` of the command
- Bootstrap package no longer stops gardener-node-init unit
- gardener-node-init unit no longer gets restarted each 30s but is a oneshot which only restarts on failure

* Address PR review feedback (nits)

Also add mutex to fake DBus to prevent flakes since the OSC controller
in GNA is running systemd commands in parallel (and in the fake DBus
they all write to the same slice in parallel)

* Consider hosts dir when pulling images via containerd

Without this, registry mirrors in the hosts dir will not be respected.

`ctr images pull` as well as the `docker.ConfigureHosts` function already ignore "not found" errors (in case the provided hosts dir does not exist), so we can safely specify it unconditionally

* containerd extractor copies file to temp dir first

and then renames it - a direct copy while a process is running will not work (`device or resource busy`)

* Address PR review feedback

* Pass entire component config to node init

This is much simpler compared to passing all potential arguments

---------

Co-authored-by: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>