Allow setting the default IPFamilies for network extension objects
This special case is necessary to allow upgrades from an older Gardener
version to succeed. In older versions, networks.extensions.gardener.cloud
objects do not contain an IPFamilies field. Thus, when upgrading a
cluster, the IPFamilies field has to be set once.

It is only allowed to set IPv4 as family because existing shoots only
use IPv4.