Do not try fetching node name on start-up
For newly created VMs, the node name is not known yet since the kubelet has to be installed first. Since GNA is the component that installs the kubelet, it doesn't make sense to try to fetch the node name on start-up.
Instead, the OSC controller tries fetching it as part of its reconciliation flow.