[node-agent] Do not try fetching node name on start-up (#8826)
* Do not try fetching node name on start-up

For newly created VMs, the node name is not known yet since the kubelet has to be installed first. Since GNA is the component that installs the kubelet, it doesn't make sense to try to fetch the node name on start-up.
Instead, the OSC controller tries fetching it as part of its reconciliation flow.

* Fix event recording

In Go, `nil` is not always `nil` - there are many topics and even official FAQ statements on this topic.
In our case, `recordEvent` takes a `runtime.Object` while the `executeUnitCommands` function (https://github.com/gardener/gardener/blob/eba989bc166232707bb1309c7df61cc6fa1ee8ca/pkg/nodeagent/controller/operatingsystemconfig/reconciler.go#L343) takes the `node` argument with type `client.Object`. When the node is not known yet (e.g., newly created VM), the `node !=nil` check in `recordEvent` results to `true` even though the value type is `nil` - but the interface type is not `nil`. This is due to the "conversion" on the way from `*corev1.Node` to `client.Object` and `runtime.Object`, finally.

Co-Authored-By: Johannes Scheerer <johannes.scheerer@sap.com>

---------

Co-authored-by: Johannes Scheerer <johannes.scheerer@sap.com>