[provider-local] Inject registry cache hosts into both OSCs
Without this, the `gardener-node-init` script will fail to extract the GNA binary from its image since this is loaded into the local registry (`localhost:5001/eu_gcr_io/...`).
When the `UseGardenerNodeAgent` feature gate gets promoted to GA (and with this locked to "always" enabled), we no longer need to inject the hosts into both OSCs. This is just needed temporary to make both scenarios work (`cloud-config-downloader` / `gardener-node-agent`), including the migration scenario.