[provider-local] Adapt to changes needed for `gardener-node-agent` (#8834)
* [provider-local] Always perform legacy 'provision' OSC reconciliation

Even when `UseGardenerNodeAgent` feature gate is enabled, the legacy OSC reconciliation logic must be performed. This is to make the migration scenario work where `cloud-config-downloader` is still running on a node. Without this, its execution script would not be updated with the new `gardener-node-agent` unit, i.e., CCD would never install `gardener-node-agent`. This would effectively break migration.

Performing both logics doesn't harm. Once the `UseGardenerNodeAgent` feature gate gets dropped, we can also cleanup this and only perform the new logic.

* Extend generic mutator interface for 'provision' OSC mutation

So far, it was only possible to mutate the 'reconcile' purpose OSC with the generic controlplane mutator.
However, with `gardener-node-agent`, also the 'provision' purpose OSC must be mutated by `provider-local` (next commit) in order to inject the registry cache host config. Otherwise, the `gardener-node-init` script will fail to extract the GNA binary from its image since this is loaded into the local registry (`localhost:5001/eu_gcr_io/...`).

* [provider-local] Inject registry cache hosts into both OSCs

Without this, the `gardener-node-init` script will fail to extract the GNA binary from its image since this is loaded into the local registry (`localhost:5001/eu_gcr_io/...`).
When the `UseGardenerNodeAgent` feature gate gets promoted to GA (and with this locked to "always" enabled), we no longer need to inject the hosts into both OSCs. This is just needed temporary to make both scenarios work (`cloud-config-downloader` / `gardener-node-agent`), including the migration scenario.

* [provider-local] Prefix GNA binary path with `/ko-app` via webhook

In the local scenario, the Gardener binaries are built with `ko`, and `ko` puts the binaries into the `/ko-app` folder in the image. This is not configurable/changeable.
Hence, provider-local must mutate the `OperatingSystemConfig`s and add the `/ko-app` prefix to the locations where the GNA binary is referenced:

- `gardener-node-init` bash script in 'provision' OSC
- `gardener-node-agent` image ref file in the 'reconcile' OSC