[node-agent] Adapt `token` controller to fetch also other tokens (`valitail`) (#8839)
* Adapt documentation

* Adapt config API

* [make generate]

* Adapt validation

* Adapt token controller business logic + integration test

* Adapt `nodeagent` component in original OSC

* Revert "Adapt `valitail`s `fetch-token.sh` script"

This reverts commit 811d43415cd8dd32b03fceed3707c2ff0a6ce58c (was part of https://github.com/gardener/gardener/pull/8827).

* Adapt `valitail` component in original OSC

Do not use `fetch-token` script when `UseGardenerNodeAgent` feature gate
is enabeld.

* Do not use `valitail` disable handling when GNA is active

gardener-node-agent takes care of disabling and removing the systemd service when it's no longer present in the operating system configuration.
Hence, the explicit disable handling/code in the unit is not needed and can be cleaned up as soon as we drop the legacy cloud-config-downloader.

* Drop legacy `promtail` handling

This was only needed when `promtail` was renamed to `valitail`, but this happened many releases ago, so we can drop this migration code now.

* Address PR review feedback