[prefactor] Introduce `authentication.gardener.cloud/__internal.KubeconfigRequest` type
We now convert the `authentication.gardener.cloud/v1alpha1.AdminKubeconfigRequest` to the `authentication.gardener.cloud/__internal.KubeconfigRequest` type when working with it. This is a prefactoring for the introduction of the new `ViewerKubeconfigRequest` API type. It will also be converted to this type internally. This enables reusing the kubeconfig creation logic (see next commits).