Enhancements for `extensions` webhook lib (#8725)
* Mutating and validating webhooks for seed

Add ability to register `MutatingWebhookConfiguration` and `ValidatingWebhookConfiguration`
objects with target `seed`. Earlier, the registration only handled mutating webhooks.

* Fix path

* Enhance webhook creation options

* Handle internal API version for webhook rules

If a webhook is primarily configured for objects of an internal API version,
the rules calculated for the webhook configuration must take all versions
("*") into consideration. Furthermore, calling `RESTMapping()` fails when
an internal version is given, so we leave it out.

* Fall back to mutating webhook

* Use `webhook.Configs` for shoot webhook configs

Make the registration for seeds and shoots consistent by also using the
`webhook.Configs` approach for target `shoot`.

* Enhance tests

* Remove obsolete CA injection

Remove CA injection in function `reconcileShootWebhookConfigs`. The data is already inserted when `AddToManager()` is invoked.

* Remove `AddToManagerSimpleOptions`

This utility will be obsolete after switching admission controllers
to certificate auto management.

* Fix path

* Address feedback I

* Add `sourceCluster` to certificate management

* Address feedback II

* Forbid using mutators and validators

* Remove hybridMutator

The HybridMutator was introduced to facilitate the InjectFunc which is however not supported any more by the controller-runtime.

https://github.com/gardener/gardener/pull/5112

* Replace `hybridValidator` with `handlerAction`

Use a `handlerAction` instead of a `mutator` and a `hybridValidator` which was always a bit confusing.

* Address feedback III