Introduce `shoots/viewerkubeconfig` subresource for requesting read-only kubeconfigs (#8870)
* Add documentation

* Create read-only RBAC resources during `Shoot` reconciliation

During shoot reconciliation, we now discover all APIs in the shoot cluster and reconcile the read-only RBAC resources. When people change the available APIs dynamically (e.g., creating or deleting CRDs or extension API servers), this will only be reflected during the next reconciliation.
This approach should be good enough for the read-only use case for now.

* Create read-only RBAC resources also for workerless shoots

* Prevent import cycle

package github.com/gardener/gardener/cmd/gardenlet
	imports github.com/gardener/gardener/cmd/gardenlet/app
	imports github.com/gardener/gardener/pkg/gardenlet/controller
	imports github.com/gardener/gardener/pkg/gardenlet/controller/shoot
	imports github.com/gardener/gardener/pkg/gardenlet/controller/shoot/care
	imports github.com/gardener/gardener/pkg/operation
	imports github.com/gardener/gardener/pkg/operation/shoot
	imports github.com/gardener/gardener/pkg/component/shootsystem
	imports github.com/gardener/gardener/pkg/operation/shoot: import cycle not allowed

* [prefactor] Introduce `authentication.gardener.cloud/__internal.KubeconfigRequest` type

We now convert the `authentication.gardener.cloud/v1alpha1.AdminKubeconfigRequest` to the `authentication.gardener.cloud/__internal.KubeconfigRequest` type when working with it. This is a prefactoring for the introduction of the new `ViewerKubeconfigRequest` API type. It will also be converted to this type internally. This enables reusing the kubeconfig creation logic (see next commits).

* [prefactor] Parametrize kubeconfig creation logic to make it reusable

* Add `authentication.gardener.cloud/v1alpha1.ViewerKubeconfig` API

* [make generate]

* Make viewer kubeconfig max duration configurable

* Add `shoots/viewerkubeconfig` subresource

* Allow project admins and viewers to create read-only kubeconfigs

* Add e2e test assertions

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback