Add support for optional SCRIPT_ROOT environment variable
This commit introduces the ability to use an optional environment variable, SCRIPT_ROOT,
in the vgopath enabled generation scripts script. When SCRIPT_ROOT is set, the script will
use this directory instead of the default $SCRIPT_DIR/.. for setting up the virtual go environment.
This enhancement provides flexibility for different runtime environments where the script's working
directory might need to be dynamically specified.