Prevent PrometheusCantScrape alerts for the etcd-druid job (#8988)
The etcd-druid scrape job in the control plane Prometheus is federating
metrics from the "cache" Prometheus in the garden namespace, which scrapes
the shared etcd-druid component.

If the --enable-backup-compaction flag is not turned on for etcd-druid,
then etcd-druid does not expose any relevant (allow listed) metrics and
the PrometheusCantScrape alert fires for all the shoots.

To prevent that, with this commit, the etcd-druid job is removed from the
scope of the generic PrometheusCantScrape alert.

This commit can be reverted if the etcd-druid job will expose at least
some metrics with any valid configuration.