Only set external traffic policy to local in local setup for multi-zone kind clusters.
The mitigation for docker desktop issues on mac (see https://github.com/gardener/gardener/pull/8972)
caused single-zone tests to fail when the istio ingress gateway pods were not scheduled on the
`gardener-local-ha-single-zone-control-plane` node.
Now, the external traffic policy is only set in scenarios when there should be istio ingress gateway
pods in all zones, i.e. in the operator setup and in the gardener multi-zone setup.