Use managed resource for shoot cluster's envoy filter resource.
To implement `kubernetes.default.svc.cluster.local` the SNI approach uses
`apiserver-proxy` component in the shoot cluster and an envoy filter on the
seed. Unfortunately, istio requires the envoy filter to be deployed in the
istio ingress gateway namespace even though it semantically belongs to the
shoot cluster control plane.
With the high-availability support in Gardener, the istio ingress gateway
responsible for shoot cluster control plane may change when it is upgraded
to be multi-zonal. Previously, the necessary envoy filter was created in the
new target namespace, but the old one was left as is. (It would be deleted
eventually when the shoot cluster would be deleted due to an owner
reference.)
This could be confusing. Therefore, this change moves the envoy filter into
a managed resource thereby ensuring that a move will automatically delete
the old resource. The result will be that only the current version will be
there for new clusters.
(This change does not handle existing envoy filters that could be deleted.
Operators are expected to delete them eventually or wait for the shoot
clusters to be deleted, which also cleans up the envoy filters.)