switch GCR -> Artifact-Registry (#8968)
GCR has been deprecated [0] in favour of Artifact-Registry.

Thus, change push-targets for OCI-Images:

- europe-docker.pkg.dev/gardener-project/snapshots for snapshots
- europe-docker.pkg.dev/gardener-project/releases for releases
- europe-docker.pkg.dev/gardener-project/publc for a combined view of
  snapshots and releases

Note that there is a corresponding change required for `branch.cfg`
residing in ref `refs/meta/ci`, which needs to be done upon merging this
Pull-Request. It is not supported by GitHub to create a Pull-Request for
this change.

[0]
https://cloud.google.com/artifact-registry/docs/transition/transition-from-gcr