[GEP-19] Migrate cache Prometheus deployment and configuration (#9128)
* Manage `ClusterRole` for prometheus instances in `prometheusoperator` component

It is shared amongst all Prometheis, so let's manage it centrally.

* Introduce `prometheus` component boilerplate

* `ServiceAccount`

`prometheus-operator` does not manage the `ServiceAccount` for the created `StatefulSet`

* `Service`

`prometheus-operator` does not manage the `Service` for the created `StatefulSet`

* `ClusterRoleBinding`

* `Prometheus` custom resource

This results in the `StatefulSet` after reconciliation of `prometheus-operator`.

* `VerticalPodAutoscaler`

We decided to no longer deploy an `Hvpa` resource since
- the HVPA controller is getting removed eventually
- it's considered no longer necessary to restrict the downscaling

* Replace rules with `PrometheusRule` resources

These rules cannot be directly associated with a component package in `pkg/component`, hence they are considered "general rules" for the cache prometheus.

* `NetworkPolicy`

* Replace `node-exporter` scrape config with `ServiceMonitor` resources

This `ServiceMonitor` cannot be directly associated with the `pkg/component/nodeexporter` package because we cannot assume that the CRDs are deployed to the shoot cluster.
Hence, it is considered "general monitor" for the cache prometheus

* Migrate `etcd-druid` scrape config into `ServiceMonitor`

* Migrate `hvpa-controller` scrape config into `ServiceMonitor`

* Migrate `kube-state-metrics` scrape config into `ServiceMonitor`

* Allow additional scrape configs

Some scrape configs might not be translatable into the CRDs (or its beyond the scope of this PR), see next commits, so we need to keep transmitting them in "raw" format. This is possible via the "additional scrape config" configuration option in the `Prometheus` CRD.

* Move `kubelet` scrape config to "additional scrape configs"

* Move `cadvisor` scrape config to "additional scrape configs"

* Cleanup no longer needed code

* Integrate cache Prometheus deployment into seed controller

* Take over existing PV by a new PVC

This way, we don't need to copy the data from one disk to another but can simply take over the existing PV (claimed by a new PVC).

I consciously decided to not invest into unit tests for the new migration.go file since
- they wouldn't yield much benefit to increase confidence in a working migration path (if we wanted to have this, we would need to invest into proper e2e tests which is even more costly and complex)
- this is temporary code that is about to get deleted soon again
- even if there's a bug that we wouldn't have caught during development/testing phase, loosing the Prometheus data is not too bad

* Documentation

* Update Prometheis configs for connection to cache Prometheus

* Address PR review feedback

* Address PR review feedback