Add container resource policies for HA VPN containers to hvpa resource of kube-apiserver. (#9173)
In the past, autoscaling of VPN side car containers was disabled in kube-apiserver. With
the removal of the old VPN solution this code disappeared.
The HA VPN solution did not influence the autoscaling of its containers in kube-apiserver.
This can lead to situations where the memory resources are so restricted by VPA that even
runc fails to start. Therefore, it is crucial to set minimum allowed values for memory so
that the containers can always start.
This change achieves exactly that.