Losen the validation of advertised addresses
The advertised addresses field in the shoot status was previously used to report addresses that can be used to reach the kube-apiserver. Since it is now going to also be used to report the service account issuer for a cluster the validation should be loosened to also allow path in the URL. In addition to that the kubeconfig returned from a adminkubeconfig request will include contexts for external, internal and managed addresses only.