Replace shoot issuer gardenlet config with a central garden secret
By design the shoot issuer domain should be centrally managed and one and the same for shoots independent from the seed. It is easier to manage the shoot issuer configuration centrally in the Garden cluster. In addition to that providing the config via gardenlet configuration file implies that the shoot issuer hostname can differ between different seeds which is not the case.