Adapt unit tests
When the feature gate is removed, the remaining unit tests and production code are adapted.
We cannot do it now because we still need the code for migration from CCD -> GNA in case somebody turned off the `UseGardenerNodeAgent` (was possible before, but is no longer possible now with GA promotion).