Reduce resource requests of istio ingress gateway and adapt autoscaling accordingly.
Due to the split of istio ingress gateways across zones in highly available seed setup,
its resource requests are slightly oversized except for very large active seed clusters.
To reduce the unnecessary resource waste this change reduces the resource requests to a
quarter (cpu) and half (memory) respectively. The general assumption is that due to its
priority the istio ingress gateway may be able to get additional cpu if available on
the node. With regards to memory, the limit is left in place with the same value and
again its priority may help not being out-of-memory killed.
As an additional measure with regards to memory, the autoscaling is extended to also
cover memory so that a scale-up can happen under memory pressure.
In addition to that, the scale-up/-down behaviour is now explicitly specified with a
fast scale-up and a slow scale-down.