Add support for migration of istio ingress namespaces.
The single-zone istio ingress gateway namespaces include the zone name as
part of their namespace name. Unfortunately, with some infrastructures
there might be inconsistencies with regards to the zone naming leading to
inconsistent istio namespace names. This is why it might be beneficial to
copy an existing istio namespace to a target namespace with a proper name.
This allows a slow migration process in the shoot maintenance windows and
is decoupled from the seed reconciliation. As a follow-up step, seed
operators may want to adapt the zones in the seed specification when all
shoot clusters have migration. This will remove the no longer needed
istio ingress gateway namespace, i.e. it will automatically clean up.