[GEP-24] Configure KAPI managed issuer & JWKS URL (#9196)
* Introduce managed issuer constants

* Add HasManagedIssuer helper func

* Validate shoot's managed issuer

* Ensure gardener-system-shoot-issuer ns creation

* Add HasManagedIssuer helper to corev1beta1

* Add ShootIssuer to gardenlet configuration

* Set managed issuer fields to kube-apiserver deployment

If a shoot requires a manage service account issuer,
the --service-account-issuer and --service-account-jwks-uri flags
of the kube-apiserver deployment will be set accordingly

* Move advertised addresses functions as part of Botanist

* Losen the validation of advertised addresses

The advertised addresses field in the shoot status was previously used to report addresses that can be used to reach the kube-apiserver. Since it is now going to also be used to report the service account issuer for a cluster the validation should be loosened to also allow path in the URL. In addition to that the kubeconfig returned from a adminkubeconfig request will include contexts for external, internal and managed addresses only.

* Replace shoot issuer gardenlet config with a central garden secret

By design the shoot issuer domain should be centrally managed and one and the same for shoots independent from the seed. It is easier to manage the shoot issuer configuration centrally in the Garden cluster. In addition to that providing the config via gardenlet configuration file implies that the shoot issuer hostname can differ between different seeds which is not the case.

* Default & require shootServiceAccountIssuerHostname func

* Call WithShootServiceAccountIssuerHostname in WithGardenFrom func

* Return error if shoot requires managed isssuer but prerequisites are not met

* Fix two linter findings

* Improve JWKS URI check & add test

* Address cosmetic suggestions

* Validate when reading hostname from secret

* Remove unnecessary conversion

* Apply review suggestion

* Address review feedback

* Extract advertised addresses names as constants

* Move ServiceAccountIssuerHostname to Shoot type

* Address feedback