Add support for highly available seed cluster, i.e. multiple availability zones.
Due to initialization order the istio ingress gateway responsible for a shoot cluster
cannot be determined when the ingress component is created. This change resolves this
problem by delaying the evaluation of the expression to a point in time when the
information is available. In other words a function is passed instead of the a
function return value.